Update patch set 1
Patch Set 1:

Needs testing (not sure how to do this non-disruptively...)

Reports on the forums for a few devices noted that dialing emergency would cause the phone to reboot. Several related Samsung devices are applying this fix - I've seen klte, jf, and d2 recently (search emergency in the search bar). This fix has been reported as working on these devices, but I have not tested it myself.

Patch-set: 1